Handle non-standard CustomRenered tags
According to the TIFF Tag Reference[1] the CustomRendered tag
specification only defines two values:
0 = Normal process
1 = Custom process

Exexif handles those two values wonderfully. The problem is that it
appears some camera manufacturers (specifically Apple) use non-standard
values[2] in the CustomRendered tag which is causing failures in exexif :(

The solution I'm offering is to confirm that the tag value is a number and
then return "Unknown (#{value})". I pulled the message format straight
from exiftool and because I'm returning the tag value I added the
is_number guard clause. I could be convinced of a different format, but
this seems sane to me.

This commit also includes a test along with an example image. Fwiw, I'm
not worried about any PII in the image exif data, I (sadly) don't live where it
was taken.

[1] http://www.awaresystems.be/imaging/tiff/tifftags/privateifd/exif/customrendered.html

[2] For example, I ran exiftool over 11K+ photos on my laptop (the
majority taken by iPhones) and about 2% had CustomRendered tags of 3, 4,
6, and 7). They definitely use 3 for HDR images but I'm not sure about
the rest.